requirements.txt
dnet==1.12
facebook-sdk==0.4.0
gevent==1.0.2
greenlet==0.4.9
pcapy==0.10.9
pymongo==3.0.3
requests==2.7.0
scapy==2.3.1
scapy-http==1.6
simplejson==3.8.0
wheel==0.24.0
